cleaned up implementation
split up Solver into different fiels/objects.
added interface to permit own implementation of a field.
